+++ deployments/test_views.py
@@ -111,20 +111,20 @@ class ProjectGetTest(APITestCase):
         # Validation Tests
         # Reached total should be provided if status is completed
         body['status'] = Statuses.COMPLETED
-        body['reached_total'] = None
+        body['reached_total'] = '' # The new framework does not allow None to be sent.
         resp = self.client.post('/api/v2/project/', body)
         self.assertEqual(resp.status_code, 400, resp.content)
 
         # Disaster Type should be provided if operation type is Long Term Operation
         body['operation_type'] = OperationTypes.PROGRAMME
-        body['dtype'] = None
+        body['dtype'] = ''
         resp = self.client.post('/api/v2/project/', body)
         self.assertEqual(resp.status_code, 400, resp.content)
 
         # Event should be provided if operation type is Emergency Operation and programme type is Multilateral
         body['operation_type'] = OperationTypes.PROGRAMME
         body['programme_type'] = ProgrammeTypes.MULTILATERAL
-        body['event'] = None
+        body['event'] = ''
         resp = self.client.post('/api/v2/project/', body)
         self.assertEqual(resp.status_code, 400, resp.content)